+++ web/js/cfm_ui.js
@@ -13,13 +13,19 @@ var seismicity_from_cache = true;
 
 var CFM_DB_tb = {
    'viewers': [
-        { 'option': 0, 'name':'CFM6_preferred','db':'CFM6_preferred_db',
+        { 'option': 0, 'name':'CFM7_preferred','db':'CFM7_preferred_db',
+          'pathname': 'research/cfm-viewer/7.0/preferred', 'port': 8104},
+        { 'option': 1, 'name':'CFM7_alternatives','db':'CFM7_alt_db',
+          'pathname': 'research/cfm-viewer/7.0/alternatives', 'port': 8106},
+        { 'option': 2, 'name':'CFM7_ruptures','db':'CFM7_rup_db',
+          'pathname': 'research/cfm-viewer/7.0/ruptures', 'port': 8108},
+        { 'option': 3, 'name':'CFM7_preferred','db':'CFM7_preferred_db',
           'pathname': 'research/cfm-viewer/6.1/preferred', 'port': 8084},
-        { 'option': 1, 'name':'CFM6_alternatives','db':'CFM6_alt_db',
+        { 'option': 4, 'name':'CFM6_alternatives','db':'CFM6_alt_db',
           'pathname': 'research/cfm-viewer/6.1/alternatives', 'port': 8086},
-        { 'option': 2, 'name':'CFM6_ruptures','db':'CFM6_rup_db',
+        { 'option': 5, 'name':'CFM6_ruptures','db':'CFM6_rup_db',
           'pathname': 'research/cfm-viewer/6.1/ruptures', 'port': 8088},
-        { 'option': 3, 'name':'CFM53_preferred','db':'CFM53_preferred_db',
+        { 'option': 6, 'name':'CFM53_preferred','db':'CFM53_preferred_db',
           'pathname': 'research/cfm-viewer/5.3/preferred', 'port': 8090}
               ]
 };