+++ web/php/declare.php
@@ -2,24 +2,24 @@
 // for Fault Objects
 $gid=0;
 $name=1;
-$alternative=2;
-$source_author=3;
-$last_update=4;
-$descriptor=5;
-$avg_strike=6;
-$avg_dip=7;
-$area_km2=8;
-$exposure=9;
-$slip_sense=10;
-$ID_comments=11;
-$USGS_ID=12;
-$fault_strand_model_description=13;
-$reference=14;
-$area=15;
-$zone=16;
-$section=17;
-$fault=18;
-$TRACE_tb_gid=19;
+#$alternative=2;
+$source_author=2;
+$last_update=3;
+#$descriptor=5;
+$avg_strike=4;
+$avg_dip=5;
+$area_km2=6;
+$exposure=7;
+$slip_sense=8;
+$ID_comments=9;
+$USGS_ID=10;
+$fault_strand_model_description=11;
+$reference=12;
+$area=13;
+$zone=14;
+$section=15;
+$fault=16;
+$TRACE_tb_gid=17;
 
 // for EQ_tb
 $quake_type_Hauksson=1;